requirements.txt
python-telegram-bot
logging
wikiapi
python-dotenv